feature: Added researh book entries for the kitchen bench feature: Material registry now looks in other mod's resources for files to load fix: Crucible ratio recipes not properly handling counts fix: Modified Dragonforging entry fix: Limestone textures being broken in OreDict context fix: Added OreDict entries for metal blocks and added mutton and rabbit to meat OreDict Entries fix: Special Armors not showing up in creative tabs fix: Carpenter custom material recipes being broken fix: Stove reproducing food out of thin air fix:

+++ src/main/java/minefantasy/mfr/tile/TileEntityBigFurnace.java
@@ -56,8 +56,8 @@ public class TileEntityBigFurnace extends TileEntityBase implements IBellowsUsea
 	private float doorAngle = 0;
 	private float prevDoorAngle = 0;
 	// HEATER
-	public float heat;
-	public float maxHeat;
+	public int heat;
+	public int maxHeat;
 	public int justShared;
 	// FURNACE
 	public int progress;
@@ -149,7 +149,7 @@ public class TileEntityBigFurnace extends TileEntityBase implements IBellowsUsea
 			if (fuel > 0) {
 				int max = (int) (maxHeat * 1.5F);
 				if (heat < max) {
-					heat += 50 * powerLevel;
+					heat += (int) (50 * powerLevel);
 				}
 
 				for (int a = 0; a < 10; a++) {
@@ -343,7 +343,7 @@ public class TileEntityBigFurnace extends TileEntityBase implements IBellowsUsea
 		}
 	}
 
-	private float getItemHeat(ItemStack itemStack) {
+	private int getItemHeat(ItemStack itemStack) {
 		return ForgeItemHandler.getForgeHeat(itemStack);
 	}
 
@@ -583,8 +583,8 @@ public class TileEntityBigFurnace extends TileEntityBase implements IBellowsUsea
 		fuel = nbt.getInteger("fuel");
 		maxFuel = nbt.getInteger("MaxFuel");
 
-		heat = nbt.getFloat("heat");
-		maxHeat = nbt.getFloat("maxHeat");
+		heat = nbt.getInteger("heat");
+		maxHeat = nbt.getInteger("maxHeat");
 
 		progress = nbt.getInteger("progress");
 		aboveType = nbt.getInteger("Level");
@@ -602,8 +602,8 @@ public class TileEntityBigFurnace extends TileEntityBase implements IBellowsUsea
 		nbt.setInteger("fuel", fuel);
 		nbt.setInteger("maxFuel", maxFuel);
 
-		nbt.setFloat("heat", heat);
-		nbt.setFloat("maxHeat", maxHeat);
+		nbt.setInteger("heat", heat);
+		nbt.setInteger("maxHeat", maxHeat);
 
 		nbt.setInteger("progress", progress);
